+++ packages/field/src/css/text-area.ts
@@ -1,0 +1,16 @@
+export default {
+  '.psds-field__text-area__container': {
+    label: 'field__text-area__container',
+
+    flex: '1'
+  },
+
+  '.psds-field__text-area': {
+    label: 'field__text-area',
+
+    border: 'none',
+    outline: 'none',
+    padding: 0,
+    width: '100%'
+  }
+}